+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 
 * Convert `iw` to `he` for files from Google Play.
+* Auto define target locale with region when locale from Google Play contains only lang code.
 
 ## 0.3.8-dev.1 - 2021-04-14